requirements.txt
acres==0.3.0
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
aspose-pdf==25.2.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
billiard==4.2.1
binaryornot==0.4.4
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
comm==0.2.2
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.1
cookiecutter==2.6.0
crispy-bootstrap5==2024.10
cron-converter==1.2.1
cron-descriptor==1.4.5
cryptography==44.0.2
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
Django==5.1.7
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-stubs==5.1.3
django-stubs-ext==5.1.3
django-timezone-field==7.1
django-timezone-utils==0.15.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
env-tools==2.4.0
et_xmlfile==2.0.0
etelemetry==0.3.1
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
fitz==0.0.1.dev2
fonttools==4.56.0
fpdf==1.7.2
fqdn==1.5.1
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.50.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.34.0
ipywidgets==8.1.5
isodate==0.6.1
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.9.0
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keyring==25.6.0
kiwisolver==1.4.8
kombu==5.4.2
looseversion==1.3.0
lxml==4.9.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
more-itertools==10.6.0
multitasking==0.0.11
mypy==1.15.0
mypy-extensions==1.0.0
narwhals==1.31.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
notebook==7.3.3
notebook_shim==0.2.4
numpy==2.2.3
openai==1.66.3
openpyxl==3.1.5
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
paramiko==3.5.1
parso==0.8.4
pathlib==1.0.1
pdfminer.six==20231228
pdfplumber==0.11.5
peewee==3.17.9
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
prometheus_client==0.21.1
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.26.1
prov==2.0.1
psutil==7.0.0
psycopg==3.2.6
psycopg-binary==3.2.6
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.28
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydot==3.0.4
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.25.4
PyNaCl==1.5.0
pyparsing==3.2.1
pypdfium2==4.30.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
python-magic==0.4.27
python-slugify==8.0.4
pytz==2025.1
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==26.3.0
RapidFuzz==3.12.2
rdflib==6.3.2
redis==5.2.1
referencing==0.36.2
reportlab==4.3.1
requests==2.32.3
requests-futures==1.0.2
RestrictedPython==8.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
scipy==1.15.2
SecretStorage==3.3.3
Send2Trash==1.8.3
setuptools==77.0.3
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
statsd==4.0.1
streamlit==1.43.2
tenacity==9.0.0
terminado==0.18.1
text-unidecode==1.3
tini==4.0.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
typer==0.15.2
types-python-dateutil==2.9.0.20241206
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
yahooquery==2.3.7
yfinance==0.2.55
